+++ TestUtils/src/main/AndroidManifest.xml
@@ -5,6 +5,6 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.microsoft.device.dualscreen.foldable_test">
+    package="com.microsoft.device.dualscreen.testutils">
 
 </manifest>
